StoreEdit: reset UI after edit-error
the public_info text is currently still stored into the DB but cut off
after the limit. Because this is considered a successful update
operation, there is no resetting of that value yet.

I don't think this is a problem, since we have a visual cue indicating
that the text is too long, which is kept until the text is shorter, or
after reloading the page of course.